refactor(message): extract WirePayload into separate wire module
Move WireType and WirePayload from tombstone.rs to new wire.rs module.
These types handle wire format discrimination for all payload types
(messages and tombstones), not just tombstones.
